+++ consts.go
@@ -1,0 +1,3 @@
+package atreugo
+
+const network = "tcp4"
